+++ app/src/main/AndroidManifest.xml
@@ -36,6 +36,9 @@
         <activity android:name=".AdminUsersActivity"
             android:exported="false"
             />
+        <activity android:name=".AdminsActivity"
+            android:exported="false"
+            />
     </application>
 
 </manifest>
