+++ app/src/main/res/layout/activity_admin.xml
@@ -32,59 +32,22 @@
         android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
         tools:layout_editor_absoluteY="417dp"
-        app:layout_constraintTop_toBottomOf="@id/demoteAdminButton"/>
-
-    <TextView
-        android:id="@+id/adminUserAdminTextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="40dp"
-        android:layout_marginTop="20dp"
-        android:padding="8dp"
-        android:text="@string/admin_users"
-        android:textColor="@color/black"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/adminToolbar" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/adminUserAdminRecyclerView"
-        android:layout_width="0dp"
-        android:layout_height="200dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/adminUserAdminTextView" />
+        app:layout_constraintTop_toBottomOf="@id/adminToolbar"/>
 
     <Button
-        android:id="@+id/deleteButtonAdmin"
+        android:id="@+id/adminsActivityButton"
         android:layout_width="150dp"
         android:layout_height="45dp"
         android:layout_margin="12dp"
-        android:background="@drawable/admindashboard_deletebutton_background"
-        android:text="@string/delete"
-        android:textColor="@color/white"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintEnd_toStartOf="@+id/demoteAdminButton"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintHorizontal_chainStyle="packed"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/adminUserAdminRecyclerView" />
-
-    <Button
-        android:id="@+id/demoteAdminButton"
-        android:layout_width="150dp"
-        android:layout_height="45dp"
-        android:layout_margin="12dp"
+        android:layout_marginEnd="128dp"
         android:background="@drawable/admindashboard_promotedemotebottom_background"
-        android:text="@string/demote"
+        android:text="@string/admins"
         android:textColor="@color/white"
         android:textSize="20sp"
         android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/deleteButtonAdmin"
-        app:layout_constraintTop_toBottomOf="@id/adminUserAdminRecyclerView" />
+        tools:layout_editor_absoluteY="417dp"
+        app:layout_constraintTop_toBottomOf="@id/usersActivityButton"/>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
