+++ app/src/main/res/layout/activity_admins.xml
@@ -1,0 +1,75 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".AdminsActivity">
+
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/adminToolbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="#424242"
+        android:backgroundTint="#FFAF52"
+        android:theme="@style/Theme.AppCompat.DayNight.DarkActionBar"
+        app:layout_constraintTop_toTopOf="parent"
+        app:navigationIcon="?attr/homeAsUpIndicator"
+        app:title="ADMIN DASHBOARD"
+        app:titleTextColor="@color/white" />
+
+    <TextView
+        android:id="@+id/adminUserAdminTextView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="40dp"
+        android:layout_marginTop="20dp"
+        android:padding="8dp"
+        android:text="@string/admin_users"
+        android:textColor="@color/black"
+        android:textSize="20sp"
+        android:textStyle="bold"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/adminToolbar" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/adminUserAdminRecyclerView"
+        android:layout_width="0dp"
+        android:layout_height="200dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/adminUserAdminTextView" />
+
+    <Button
+        android:id="@+id/deleteButtonAdmin"
+        android:layout_width="150dp"
+        android:layout_height="45dp"
+        android:layout_margin="12dp"
+        android:background="@drawable/admindashboard_deletebutton_background"
+        android:text="@string/delete"
+        android:textColor="@color/white"
+        android:textSize="20sp"
+        android:textStyle="bold"
+        app:layout_constraintEnd_toStartOf="@+id/demoteAdminButton"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintHorizontal_chainStyle="packed"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/adminUserAdminRecyclerView" />
+
+    <Button
+        android:id="@+id/demoteAdminButton"
+        android:layout_width="150dp"
+        android:layout_height="45dp"
+        android:layout_margin="12dp"
+        android:background="@drawable/admindashboard_promotedemotebottom_background"
+        android:text="@string/demote"
+        android:textColor="@color/white"
+        android:textSize="20sp"
+        android:textStyle="bold"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toEndOf="@+id/deleteButtonAdmin"
+        app:layout_constraintTop_toBottomOf="@id/adminUserAdminRecyclerView" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
